Changing the structure of the server
-	Moving files into API folder to separate layers of Controller and Models(DB)

-	Combine validation of users data for login and registration

-	Deleting unnecessary files
