+++ api/routes/index.js
@@ -1,0 +1,5 @@
+const user = require('./user');
+
+module.exports = {
+  user
+};
