+++ api/services/registrationUser.js
@@ -1,4 +1,4 @@
-const db = require('../database');
+const db = require('../../database');
 const bcrypt = require('bcryptjs');
 const ErrorHandler = require('../services/error.handler');
