+++ api/validation/user.validation.js
@@ -1,16 +1,14 @@
 const Joi = require('@hapi/joi');
 
-module.exports = Joi.object().keys({
+const keys = {
   firstName: Joi.string()
     .min(1)
-    .max(20)
-    .required(),
+    .max(20),
   lastName: Joi.string()
     .min(1)
-    .max(20)
-    .required(),
+    .max(20),
   email: Joi.string()
-    .max(64)
+    .max(255)
     .email({ minDomainSegments: 2 })
     .required(),
   password: Joi.string()
@@ -19,4 +17,13 @@ module.exports = Joi.object().keys({
     .regex(/^[a-zA-Z0-9!@#$%^&*?]/)
     .required(),
   passwordConfirm: Joi.ref('password'),
-});
+};
+
+const registrationSchema = Joi.object().keys(keys);
+const loginSchema = Joi.object().keys(keys).optionalKeys('firstName', 'lastName', 'passwordConfirm');
+
+module.exports = {
+  loginSchema,
+  registrationSchema,
+  
+};
