--- routes/index.js
@@ -1,7 +1,0 @@
-const authRouter = {};
-const registrationRouter = require('./registration.router');
-
-module.exports = {
-  authRouter,
-  registrationRouter,
-};
